nmr_flask_sparql_frontend/requirements.txt
flask
requests
python-dotenv
python-keycloak